+++ custom_components/solarcharger/config_subentry_flow.py
@@ -141,12 +141,7 @@ class AddChargerSubEntryFlowHandler(ConfigSubentryFlow):
         data: dict[str, Any] = {
             OPTION_CHARGER_DEVICE_NAME: device_name,
         }
-        reset_api_entities(
-            config_entry,
-            subentry_unique_id,
-            data,
-            reset_all_entities=True,
-        )
+        reset_api_entities(config_entry, subentry_unique_id, data)
 
         self.hass.config_entries.async_update_entry(
             config_entry,