+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v0.2beta8 2026-01-02
+- Added switch to enable/disable plug in trigger.
+- Added switch to enable/disable sun elevation trigger.
+
 ## v0.2beta7 2025-12-31
 - Support for Tesla MQTT BLE API with initial testing ok.
 - Added support for Tesla Fleet and Tessie APIs but not tested.